requirements.txt
streamlit
streamlit-chat
python-dotenv
langchain-core
langchain-google-genai
google-generativeai
sentence-transformers
pinecone-client